1_sportprog/30/requirements.txt
codeforces_api==0.3.4
CodeforcesApiPy==2.0.10
python-decouple==3.6
